+++ R/dbConnect.R
@@ -26,3 +26,29 @@ connect_online <- function(dbname = "gbif4crest_02", port = 5432, host = "gbif4c
     )
     return(db)
 }
+
+
+
+#' Connect to the gbif4crest calibration database
+#'
+#' Connect to the gbif4crest_02 database using a local SQLite3 copy.
+#'
+#' @param dbname The complete path to the SQLite3 file. The name should end by '.sqlite3'
+#' @return An active connection to a database
+#' @export
+#' @seealso The SQLite3 database can be downloaded from \url{https://figshare.com/articles/GBIF_for_CREST_database/6743207}.
+#' @examples
+#' \dontrun{
+#'   db <- connect_online()
+#' }
+#'
+connect_local_sqlite3 <- function(dbname = "gbif4crest_02.sqlite3") {
+    if (!requireNamespace("RSQLite", quietly = TRUE)) {
+        stop("The package 'RSQLite' is required. Use install.packages('RSQLite').\n")
+    }
+    db <- DBI::dbConnect(
+      drv = RSQLite::SQLite(),
+      dbname
+    )
+    return(db)
+}